requirements.txt
python-dotenv
selenium
gologin
webdriver-manager
pyotp
pydantic_settings